2020 08 19

+++ src/main/java/com/xyc/userc/dao/MobileMapper.java
@@ -13,7 +13,7 @@ public interface MobileMapper
     public String selectMesCodeByMobile(String mobile);
 
     public void insertMesCode(@Param("mobile")String mobile, @Param("mesCode")String mesCode,
-                              @Param("status")Byte status, @Param("gmtCreate")Date gmtCreate,
+                              @Param("status")int status, @Param("gmtCreate")Date gmtCreate,
                               @Param("gmtModified")Date gmtModified);
 
     public void updateMesCodeStatus(@Param("mobile")String mobile, @Param("status")Byte status,
